+++ app/services/integrations/__init__.py
@@ -4,6 +4,8 @@ from app.services.integrations.github import GitHubClient
 from app.services.integrations.github_parser import GitHubDataParser
 from app.services.integrations.github_sync import GitHubSyncService
 from app.services.integrations.leetcode import LeetCodeClient
+from app.services.integrations.leetcode_parser import LeetCodeDataParser
+from app.services.integrations.leetcode_sync import LeetCodeSyncService
 
 __all__ = [
     "BasePlatformClient",
@@ -13,6 +15,8 @@ __all__ = [
     "GitHubDataParser",
     "GitHubSyncService",
     "LeetCodeClient",
+    "LeetCodeDataParser",
+    "LeetCodeSyncService",
 ]